requirements.txt
discord.py>=2.3.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
python-dotenv>=1.0.0
schedule>=1.2.0